// packages/log-core/src/core/types/types-messages.ts
export type MessageRole = "system" | "user" | "assistant" | "tool";

export interface Message {
	role: MessageRole;
	content: string | null;
	tool_call_id?: string;
	tool_calls?: ToolCall[];
	name?: string;
	timestamp?: number;
}
// ...
export interface ToolCall {
	id: string;
	name: string;
	arguments: string;
}